+++ spec/client_spec.rb
@@ -36,13 +36,11 @@ RSpec.describe MitakeSms::Client do
         stubs.post('SmSend') do |env|
           expect(env.url.path).to eq('/SmSend')
           # Check for hash parameters
-          expect(env.body).to eq({
-            username: 'test_username',
-            password: 'test_password',
-            dstaddr: to,
-            smbody: 'Test message',
-            CharsetURL: 'UTF8'
-          })
+          expect(env.body[:username]).to eq('test_username')
+          expect(env.body[:password]).to eq('test_password')
+          expect(env.body[:dstaddr]).to eq(to)
+          expect(env.body[:smbody]).to eq(URI.encode_www_form_component('Test message'))
+          expect(env.body[:CharsetURL]).to eq('UTF8')
 
           [
             200,
@@ -87,12 +85,11 @@ RSpec.describe MitakeSms::Client do
         stubs.post('SmBulkSend') do |env|
           expect(env.url.path).to eq('/SmBulkSend')
           # Check for hash parameters
-          expect(env.body).to eq({
-            username: 'test_username',
-            password: 'test_password',
-            smbody: "0912345678:Message 1\n0922333444:Message 2",
-            Encoding_PostIn: 'UTF8'
-          })
+          expect(env.body[:username]).to eq('test_username')
+          expect(env.body[:password]).to eq('test_password')
+          expect(env.body[:smbody]).to include(URI.encode_www_form_component('Message 1'))
+          expect(env.body[:smbody]).to include(URI.encode_www_form_component('Message 2'))
+          expect(env.body[:Encoding_PostIn]).to eq('UTF8')
 
           [
             200,
@@ -124,12 +121,11 @@ RSpec.describe MitakeSms::Client do
       before do
         stubs.post('SmBulkSend') do |env|
           expect(env.url.path).to eq('/SmBulkSend')
-          expect(env.body).to eq({
-            username: 'test_username',
-            password: 'test_password',
-            smbody: "0912345678:Message 1\n0922333444:Message 2",
-            Encoding_PostIn: 'UTF8'
-          })
+          expect(env.body[:username]).to eq('test_username')
+          expect(env.body[:password]).to eq('test_password')
+          expect(env.body[:smbody]).to include(URI.encode_www_form_component('Message 1'))
+          expect(env.body[:smbody]).to include(URI.encode_www_form_component('Message 2'))
+          expect(env.body[:Encoding_PostIn]).to eq('UTF8')
 
           [
             200,
@@ -160,16 +156,22 @@ RSpec.describe MitakeSms::Client do
       end
 
       before do
-        # First batch stub
+        # Set up counter to track which batch is being processed
+        batch_counter = 0
+        
+        # Stub for both batches
         stubs.post('SmBulkSend') do |env|
           expect(env.url.path).to eq('/SmBulkSend')
-          if env.body[:smbody].include?('Message 1')
-            expect(env.body).to eq({
-              username: 'test_username',
-              password: 'test_password',
-              smbody: "0912345678:Message 1\n0922333444:Message 2",
-              Encoding_PostIn: 'UTF8'
-            })
+          
+          batch_counter += 1
+          
+          if batch_counter == 1
+            # First batch should contain Message 1 and Message 2
+            expect(env.body[:username]).to eq('test_username')
+            expect(env.body[:password]).to eq('test_password')
+            expect(env.body[:smbody]).to include(URI.encode_www_form_component('Message 1'))
+            expect(env.body[:smbody]).to include(URI.encode_www_form_component('Message 2'))
+            expect(env.body[:Encoding_PostIn]).to eq('UTF8')
 
             [
               200,
@@ -177,12 +179,12 @@ RSpec.describe MitakeSms::Client do
               "statuscode=1\nmsgid=1234567890\nAccountPoint=98"
             ]
           else
-            expect(env.body).to eq({
-              username: 'test_username',
-              password: 'test_password',
-              smbody: "0933555666:Message 3\n0944666777:Message 4",
-              Encoding_PostIn: 'UTF8'
-            })
+            # Second batch should contain Message 3 and Message 4
+            expect(env.body[:username]).to eq('test_username')
+            expect(env.body[:password]).to eq('test_password')
+            expect(env.body[:smbody]).to include(URI.encode_www_form_component('Message 3'))
+            expect(env.body[:smbody]).to include(URI.encode_www_form_component('Message 4'))
+            expect(env.body[:Encoding_PostIn]).to eq('UTF8')
 
             [
               200,